Converted everything to single precision.

+++ color_tools.c
@@ -3,31 +3,31 @@
 
 #include "color_tools.h"
 
-double Al[3][3] = {
+float Al[3][3] = {
     {0.1840, 0.0179, 0.0048}, 
     {0.0876, 0.0118, 0.0018}, 
     {0.0005, 0.0012, 0.0159}
 };
-double Ar[3][3] = {
+float Ar[3][3] = {
     {0.0153, 0.1092, 0.1171}, 
     {0.0176, 0.3088, 0.0777}, 
     {0.0201, 0.1016, 0.6546}
 };
-double CIEmatrix[3][3] = {
+float CIEmatrix[3][3] = {
     {0.4243, 0.3105, 0.1657}, 
     {0.2492, 0.6419, 0.1089}, 
     {0.0265, 0.1225, 0.8614}
 };
 
-double Xn = 111.144;
-double Yn = 100.0;
-double Zn = 35.201;
+float Xn = 111.144;
+float Yn = 100.0;
+float Zn = 35.201;
 
-double CIEmatrixL[3][3];
-double CIEmatrixR[3][3];
+float CIEmatrixL[3][3];
+float CIEmatrixR[3][3];
 
 
-void mult_matrices(double a[][3], double b[3], double result[3])
+void mult_matrices(float a[][3], float b[3], float result[3])
 {
     for (int i=0; i<3; i++) {
         result[i] = 0;
@@ -44,7 +44,7 @@ void mult_matrices(double a[][3], double b[3], double result[3])
    }
 }
 
-void print_matrix(double a[3])
+void print_matrix(float a[3])
 {
    int i; 
    for (i=0; i<3; i++)
@@ -64,11 +64,11 @@ left and right eye views
 
 void CIEmatricesLandR()
 {
-    double CIEleftconstant[3];
-    double CIErightconstant[3];
-    double CIEmatrixrightconstant[3];
-    double CIEmatrixleftconstant[3];
-    double RGBwhite[3] = {255, 255, 255};
+    float CIEleftconstant[3];
+    float CIErightconstant[3];
+    float CIEmatrixrightconstant[3];
+    float CIEmatrixleftconstant[3];
+    float RGBwhite[3] = {255, 255, 255};
 
     mult_matrices(Al, RGBwhite, CIEleftconstant);
     mult_matrices(Ar, RGBwhite, CIErightconstant);
@@ -78,7 +78,7 @@ void CIEmatricesLandR()
 
     //compute normalization values n2, ..., n6 so that that whites agree in
     //CIE color space
-    double n[6];
+    float n[6];
     for(int i=0; i<3; i++){
         n[i] = CIEleftconstant[i] / CIEmatrixleftconstant[i];
     }
@@ -93,7 +93,7 @@ void CIEmatricesLandR()
 }
 
 //element_wise_multiply
-void element_wise_multiply(double matrix[][3], double vector[3], double result[][3])
+void element_wise_multiply(float matrix[][3], float vector[3], float result[][3])
 {
     for (int i=0; i<3; i++){
         for(int j=0; j<3; j++){
@@ -102,10 +102,10 @@ void element_wise_multiply(double matrix[][3], double vector[3], double result[]
     }
 }
 
-double cielabf(double U, double Un)
+float cielabf(float U, float Un)
 {
-    double returnval;
-    double result = U/Un;
+    float returnval;
+    float result = U/Un;
     //printf("result = %f\n", result);
     if(result > .008856) {
         returnval = pow(result,1.0/3);
@@ -116,28 +116,28 @@ double cielabf(double U, double Un)
     return returnval;
 }
 
-double Lstar(double x, double y, double z) 
+float Lstar(float x, float y, float z) 
 {
-    double r = (116.0 * cielabf(y, Yn) - 16.0);
+    float r = (116.0 * cielabf(y, Yn) - 16.0);
     //printf("Lstar is %f\n", r);
     return r;
 }
 
-double Astar(double x, double y, double z)
+float Astar(float x, float y, float z)
 {
-    double r = 500.0 * (cielabf(x, Xn) - cielabf(y, Yn));
+    float r = 500.0 * (cielabf(x, Xn) - cielabf(y, Yn));
     //printf("Astar is  %f\n", r);
     return r;
 }
 
-double Bstar(double x, double y, double z) 
+float Bstar(float x, float y, float z) 
 {
     return (200.0 * (cielabf(y, Yn) - cielabf(z, Zn)));
 }
 
-double Lstarleft(double rgb[3])
+float Lstarleft(float rgb[3])
 {
-    double CIEmatrixLeft[3];
+    float CIEmatrixLeft[3];
 
     mult_matrices(CIEmatrixL, rgb, CIEmatrixLeft);
     /*for (int i=0; i<3; i++){
@@ -147,34 +147,34 @@ double Lstarleft(double rgb[3])
         printf("\n");
     }*/
     //printf("RGB: %f,%f,%f\n", rgb[0], rgb[1], rgb[2]);
-    double test = Lstar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]);
+    float test = Lstar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]);
     //printf("Lstarleft is %f\n", test);
     return test;
 
 }
 
-double Astarleft(double rgb[3])
+float Astarleft(float rgb[3])
 {
-    double CIEmatrixLeft[3];
+    float CIEmatrixLeft[3];
     mult_matrices(CIEmatrixL, rgb, CIEmatrixLeft);
-    double test = Astar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]); 
+    float test = Astar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]); 
     //printf("Astarleft is %f\n", test);
     return test;
 
 }
 
-double Bstarleft(double rgb[3])
+float Bstarleft(float rgb[3])
 {
-    double CIEmatrixLeft[3];
+    float CIEmatrixLeft[3];
     mult_matrices(CIEmatrixL, rgb, CIEmatrixLeft);
-    double test = Bstar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]);
+    float test = Bstar(CIEmatrixLeft[0], CIEmatrixLeft[1], CIEmatrixLeft[2]);
     //printf("Bstarleft is %f\n", test);
     return test;
 }
 
-double Lstarright(double rgb[3])
+float Lstarright(float rgb[3])
 {
-    double CIEmatrixRight[3];
+    float CIEmatrixRight[3];
     mult_matrices(CIEmatrixR, rgb, CIEmatrixRight);
     /*for (int i=0; i<3; i++){
         for(int j=0; j<3; j++) {
@@ -183,92 +183,92 @@ double Lstarright(double rgb[3])
         printf("\n");
     }*/
 
-    double test = Lstar(CIEmatrixRight[0], CIEmatrixRight[1], CIEmatrixRight[2]);
+    float test = Lstar(CIEmatrixRight[0], CIEmatrixRight[1], CIEmatrixRight[2]);
     //printf("Lstarright is %f\n", test);
     return test;
 }
 
-double Astarright(double rgb[3])
+float Astarright(float rgb[3])
 {
-    double CIEmatrixRight[3];
+    float CIEmatrixRight[3];
     mult_matrices(CIEmatrixR, rgb, CIEmatrixRight);
     return Astar(CIEmatrixRight[0], CIEmatrixRight[1], CIEmatrixRight[2]);
 
 }
 
-double Bstarright(double rgb[3])
+float Bstarright(float rgb[3])
 {
-    double CIEmatrixRight[3];
+    float CIEmatrixRight[3];
     mult_matrices(CIEmatrixR, rgb, CIEmatrixRight);
     return Bstar(CIEmatrixRight[0], CIEmatrixRight[1], CIEmatrixRight[2]);
 
 }
 
-void CIE(double A[][3], double rgb[3], double result[3])
+void CIE(float A[][3], float rgb[3], float result[3])
 {
     mult_matrices(A, rgb, result);
 }
 
-double CLstarLeft(double rgb[3]) 
+float CLstarLeft(float rgb[3]) 
 {
-    double CIELeft[3];
+    float CIELeft[3];
     CIE(Al, rgb, CIELeft);
     return Lstar(CIELeft[0], CIELeft[1], CIELeft[2]);
 }
 
-double CAstarLeft(double rgb[3])
+float CAstarLeft(float rgb[3])
 {
-    double CIELeft[3];
+    float CIELeft[3];
     CIE(Al, rgb, CIELeft);
     return Astar(CIELeft[0], CIELeft[1], CIELeft[2]);
 }
 
-double CBstarLeft(double rgb[3])
+float CBstarLeft(float rgb[3])
 {
-    double CIELeft[3];
+    float CIELeft[3];
     CIE(Al, rgb, CIELeft);
     //printf("CIELeft is %f, %f, %f\n", CIELeft[0], CIELeft[1], CIELeft[2]);
     return Bstar(CIELeft[0], CIELeft[1], CIELeft[2]);
 }
 
-double CLstarRight(double rgb[3]) 
+float CLstarRight(float rgb[3]) 
 {
-    double CIEright[3];
+    float CIEright[3];
     CIE(Ar, rgb, CIEright);
     return Lstar(CIEright[0], CIEright[1], CIEright[2]);
 }
 
-double CAstarRight(double rgb[3])
+float CAstarRight(float rgb[3])
 {
-    double CIEright[3];
+    float CIEright[3];
     CIE(Ar, rgb, CIEright);
     return Astar(CIEright[0], CIEright[1], CIEright[2]);
 }
 
-double CBstarRight(double rgb[3])
+float CBstarRight(float rgb[3])
 {
-    double CIEright[3];
+    float CIEright[3];
     CIE(Ar, rgb, CIEright);
     return Bstar(CIEright[0], CIEright[1], CIEright[2]);
 }
 
-double LCD_Lstar(double rgb[3])
+float LCD_Lstar(float rgb[3])
 {
-    double lcd_CIE[3];
+    float lcd_CIE[3];
     CIE(CIEmatrix, rgb, lcd_CIE);
     return Lstar(lcd_CIE[0], lcd_CIE[1], lcd_CIE[2]);
 }
 
-double LCD_Astar(double rgb[3])
+float LCD_Astar(float rgb[3])
 {
-    double lcd_CIE[3];
+    float lcd_CIE[3];
     CIE(CIEmatrix, rgb, lcd_CIE);
     return Astar(lcd_CIE[0], lcd_CIE[1], lcd_CIE[2]);
 }
 
-double LCD_Bstar(double rgb[3])
+float LCD_Bstar(float rgb[3])
 {
-    double lcd_CIE[3];
+    float lcd_CIE[3];
     CIE(CIEmatrix, rgb, lcd_CIE);
     return Bstar(lcd_CIE[0], lcd_CIE[1], lcd_CIE[2]);
 }
